Add clickmod, automatic env-specific docgen modification
This now supports automatically converting references in docstrings to a suitable form both for the CLI and the HTML documentation.
In order to use it, instead of importing click via "import click", instead use "from clickmod import click". This makes the necessary adjustments to click before handing it to you.
(If clickmod.py is in the same directory you may need to write "from .clickmod import click" instead. See Python modules and packages for more information.)

Caveat: In order to support cross-referencing for groups and commands, you will need to also set the html_prefix property of clickmod.clickmod_settings.
This unfortunately needs to be done wherever you import click in the aforementioned way. For example:

    from .clickmod import click, clickmod_settings
    clickmod_settings.html_prefix = "cli-"

It may be possible to change this but I may scrap the entire necessity to use direct HTML links if we decide that I should modify sphinx-click to add custom labels for groups and commands, so I'm leaving that alone for now.