Works with the modifications to sphinx-click to generate a "complete" group page
Next up is to properly replace the pip-installed sphinx-click library with the local modified version (rather than use "sphinx-click2").
Then I'll separate the regexes on clickmod.py and add a new one for a general reference, since they all work the same now and the information of what type of reference it is is no longer necessary.
Will also consider seeing if local modules can supersede global modules in imports to present the option of the invisible click wrapping, even though I'm not sure it's a good idea in general.